# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BattleJoinInput {
  /**
     * @minLength 1
     * @maxLength 18
     */
  name: string;
  /**
     * @minLength 16
     * @maxLength 128
     * @pattern ^[A-Za-z0-9._:-]+$
     */
  requestId?: string;
}

export interface BattleAccessInput { [key: string]: unknown }

export type BattleShotInputFireMode = typeof BattleShotInputFireMode[keyof typeof BattleShotInputFireMode];


export const BattleShotInputFireMode = {
  primary: 'primary',
  throw: 'throw',
} as const;

export interface BattleShotInput {
  /**
     * @minimum 0
     * @maximum 9
     */
  markerId: number;
  firedAt: number;
  /**
     * @minLength 1
     * @maxLength 32
     */
  weaponId: string;
  fireMode: BattleShotInputFireMode;
}

export type NetworkBattleShotInputFireMode = typeof NetworkBattleShotInputFireMode[keyof typeof NetworkBattleShotInputFireMode];


export const NetworkBattleShotInputFireMode = {
  primary: 'primary',
  throw: 'throw',
} as const;

export interface NetworkBattleShotInput {
  /**
     * @minLength 16
     * @maxLength 128
     * @pattern ^[A-Za-z0-9._:-]+$
     */
  shotId: string;
  /**
     * @minLength 1
     * @maxLength 128
     */
  targetPlayerId: string;
  firedAt: number;
  /**
     * @minLength 1
     * @maxLength 32
     */
  weaponId: string;
  fireMode: NetworkBattleShotInputFireMode;
}

export type BattleEffectKind = typeof BattleEffectKind[keyof typeof BattleEffectKind];


export const BattleEffectKind = {
  frag: 'frag',
  flashbang: 'flashbang',
  smoke: 'smoke',
} as const;

export interface BattleEffect {
  id: string;
  kind: BattleEffectKind;
  expiresAt: number;
}

export interface BattlePlayer {
  id: string;
  name: string;
  markerId: number;
  markerColor: string;
  lives: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  hp: number;
  alive: boolean;
  respawnAt: number;
  isHost: boolean;
  connected: boolean;
}

export type BattleRoomStatus = typeof BattleRoomStatus[keyof typeof BattleRoomStatus];


export const BattleRoomStatus = {
  lobby: 'lobby',
  active: 'active',
  finished: 'finished',
} as const;

export interface BattleRoom {
  code: string;
  status: BattleRoomStatus;
  winnerPlayerId: string | null;
  draw: boolean;
  players: BattlePlayer[];
  effects: BattleEffect[];
  maxPlayers: number;
  updatedAt: number;
}

export interface BattleSession {
  playerId: string;
  sessionToken: string;
  room: BattleRoom;
}

export interface BattleShotResult {
  accepted: boolean;
  reason: string;
  shooterId: string;
  targetId: string;
  /** @minimum 0 */
  damage: number;
  eliminated: boolean;
  room: BattleRoom;
}

export type GetBattleStateParams = {
/**
 * @pattern ^[A-Z0-9]{6}$
 */
code: string;
};