# lib/api-zod/src/generated/types/networkBattleShotInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NetworkBattleShotInputFireMode } from './networkBattleShotInputFireMode';

export interface NetworkBattleShotInput {
  /**
     * @minLength 16
     * @maxLength 128
     * @pattern ^[A-Za-z0-9._:-]+$
     */
  shotId: string;
  /**
     * @minLength 1
     * @maxLength 128
     */
  targetPlayerId: string;
  firedAt: number;
  /**
     * @minLength 1
     * @maxLength 32
     */
  weaponId: string;
  fireMode: NetworkBattleShotInputFireMode;
}